refactor: 검색 관련 타입과 유틸 함수 공통화
